use ip addresses instead of hostnames for allow lists
In the Apache 2.4 beaches of some modules, "Require ip" is used, and that only works on ip addresses.

The choice is to either set the default attributes to IPs, change the modules to use "Require host", or split the allow lists into an ipaddress and hostname version so the user can specify both.

I've just changed the defaults to use IPs.